+++ lib/extension/string_extension.dart
@@ -1,0 +1,21 @@
+extension StringExtension on String {
+  bool containsAny(List<String> values) {
+    for (final value in values) {
+      if (contains(value)) {
+        return true;
+      }
+    }
+
+    return false;
+  }
+
+  bool containsAll(List<String> values) {
+    for (final value in values) {
+      if (!contains(value)) {
+        return false;
+      }
+    }
+
+    return true;
+  }
+}