+++ lib/ui/emulator/rom_manager.g.dart
@@ -7,17 +7,21 @@ part of 'rom_manager.dart';
 // **************************************************************************
 
 RomInfo _$RomInfoFromJson(Map<String, dynamic> json) => RomInfo(
-  name: json['name'] as String,
-  path: json['path'] as String,
-  hash: json['hash'] as String,
-  slot: (json['slot'] as num?)?.toInt(),
+  name: json['name'] as String?,
+  path: json['path'] as String?,
+  hash: json['hash'] as String?,
+  romHash: json['romHash'] as String?,
+  chrHash: json['chrHash'] as String?,
+  prgHash: json['prgHash'] as String?,
 );
 
 Map<String, dynamic> _$RomInfoToJson(RomInfo instance) => <String, dynamic>{
   'name': instance.name,
   'path': instance.path,
   'hash': instance.hash,
-  'slot': instance.slot,
+  'romHash': instance.romHash,
+  'chrHash': instance.chrHash,
+  'prgHash': instance.prgHash,
 };
 
 // **************************************************************************