+++ lib/ui/save_states/save_states_screen.dart
@@ -73,7 +73,7 @@ class SaveStatesScreen extends HookConsumerWidget {
     return NesdScaffold(
       appBar: AppBar(
         title: Text(
-          'Save States - ${p.basenameWithoutExtension(romInfo.name)}',
+          'Save States - ${p.basenameWithoutExtension(romInfo.name ?? '')}',
           style: TextStyle(
             color: Theme.of(context).colorScheme.primary,
             fontWeight: FontWeight.bold,
@@ -100,7 +100,7 @@ class SaveStatesScreen extends HookConsumerWidget {
                     romTileData: romTileData,
                     onPressed: () {
                       nesController.loadRom(
-                        romTileData.romInfo.path,
+                        romTileData.romInfo.path ?? '',
                         state: romTileData.state,
                       );